pyproject.toml
[project]
name = "reinforcement_learning_taxi"
version = "0.1.0"
description = "Reinforcement Learning comparison: Q-Learning vs DQN on Taxi-v3"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "gymnasium>=0.29.1",
    "stable-baselines3>=2.3.0",
    "sb3-contrib>=2.3.0",
    "torch>=2.0.0",
    "numpy>=1.24.0",
    "matplotlib>=3.7.0",
    "seaborn>=0.12.0",
    "pandas>=2.0.0",
    "pyyaml>=6.0",
    "tensorboard>=2.13.0",
    "tqdm>=4.65.0",
    "imageio>=2.31.0",
    "imageio-ffmpeg>=0.4.8",
    "scipy>=1.10.0",
    "jupyter>=1.0.0",
    "ipywidgets>=8.0.0",
    "pygame>=2.6.1",
    "moviepy>=2.2.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/reinforcement_learning_taxi"]

[dependency-groups]
dev = [
    "pytest>=7.0.0",
    "ruff>=0.1.0",
    "mypy>=1.0.0",
]

[tool.ruff]
line-length = 120
target-version = "py310"
select = ["E", "F", "I"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
